requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
pypdf
transformers
sentence-transformers
huggingface-hub
